+++ pysofe/elements/__init__.py
@@ -4,4 +4,4 @@ Provides the data structure for finite elements.
 
 import simple
 
-from simple.lagrange import P1
+from simple.lagrange import P1, P2