+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Unreleased
 Added
 -----
 
-- Linear Lagrange element P1
+- Lagrange elements P1, P2
 - Basic mesh data structures and uniform refinement method
 - Gauss-Legendre quadrature for simplicial domains
 - FE space class